+++ packages/Angular/Explorer/compare-records/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log - @memberjunction/ng-compare-records
 
+## 2.57.0
+
+### Patch Changes
+
+- Updated dependencies [0ba485f]
+  - @memberjunction/core@2.57.0
+  - @memberjunction/core-entities@2.57.0
+
 ## 2.56.0
 
 ### Patch Changes