+++ packages/Angular/Explorer/form-toolbar/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log - @memberjunction/ng-form-toolbar
 
+## 2.57.0
+
+### Patch Changes
+
+- Updated dependencies [0ba485f]
+  - @memberjunction/core@2.57.0
+  - @memberjunction/global@2.57.0
+  - @memberjunction/ng-ask-skip@2.57.0
+  - @memberjunction/ng-base-forms@2.57.0
+  - @memberjunction/ng-record-changes@2.57.0
+  - @memberjunction/ng-shared@2.57.0
+  - @memberjunction/ng-container-directives@2.57.0
+
 ## 2.56.0
 
 ### Patch Changes