+++ packages/Angular/Explorer/shared/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log - @memberjunction/ng-shared
 
+## 2.57.0
+
+### Patch Changes
+
+- Updated dependencies [0ba485f]
+  - @memberjunction/core@2.57.0
+  - @memberjunction/core-entities@2.57.0
+  - @memberjunction/ng-base-types@2.57.0
+  - @memberjunction/ng-notifications@2.57.0
+  - @memberjunction/graphql-dataprovider@2.57.0
+
 ## 2.56.0
 
 ### Patch Changes