+++ packages/Angular/Explorer/user-view-properties/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log - @memberjunction/ng-user-view-properties
 
+## 2.57.0
+
+### Patch Changes
+
+- Updated dependencies [0ba485f]
+  - @memberjunction/core@2.57.0
+  - @memberjunction/core-entities@2.57.0
+  - @memberjunction/global@2.57.0
+  - @memberjunction/ng-base-forms@2.57.0
+  - @memberjunction/ng-shared@2.57.0
+  - @memberjunction/ng-find-record@2.57.0
+  - @memberjunction/ng-resource-permissions@2.57.0
+
 ## 2.56.0
 
 ### Patch Changes