+++ packages/Angular/Generic/action-gallery/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @memberjunction/ng-action-gallery
 
+## 2.57.0
+
+### Patch Changes
+
+- Updated dependencies [0ba485f]
+  - @memberjunction/core@2.57.0
+  - @memberjunction/core-entities@2.57.0
+  - @memberjunction/ng-ai-test-harness@2.57.0
+  - @memberjunction/ng-container-directives@2.57.0
+
 ## 2.56.0
 
 ### Patch Changes