+++ packages/Communication/base-types/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log - @memberjunction/communication-types
 
+## 2.57.0
+
+### Patch Changes
+
+- Updated dependencies [0ba485f]
+  - @memberjunction/core@2.57.0
+  - @memberjunction/core-entities@2.57.0
+  - @memberjunction/global@2.57.0
+  - @memberjunction/templates-base-types@2.57.0
+
 ## 2.56.0
 
 ### Patch Changes