+++ packages/Communication/engine/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log - @memberjunction/communication-engine
 
+## 2.57.0
+
+### Patch Changes
+
+- Updated dependencies [0ba485f]
+  - @memberjunction/core@2.57.0
+  - @memberjunction/core-entities@2.57.0
+  - @memberjunction/global@2.57.0
+  - @memberjunction/communication-types@2.57.0
+  - @memberjunction/templates@2.57.0
+
 ## 2.56.0
 
 ### Patch Changes