+++ packages/Communication/entity-comm-client/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log - @memberjunction/entity-communications-client
 
+## 2.57.0
+
+### Patch Changes
+
+- Updated dependencies [0ba485f]
+  - @memberjunction/core@2.57.0
+  - @memberjunction/core-entities@2.57.0
+  - @memberjunction/global@2.57.0
+  - @memberjunction/entity-communications-base@2.57.0
+  - @memberjunction/graphql-dataprovider@2.57.0
+
 ## 2.56.0
 
 ### Patch Changes