+++ packages/Communication/providers/MSGraph/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log - @memberjunction/communication-ms-graph
 
+## 2.57.0
+
+### Patch Changes
+
+- Updated dependencies [0ba485f]
+  - @memberjunction/core@2.57.0
+  - @memberjunction/core-entities@2.57.0
+  - @memberjunction/global@2.57.0
+  - @memberjunction/sqlserver-dataprovider@2.57.0
+  - @memberjunction/aiengine@2.57.0
+  - @memberjunction/communication-types@2.57.0
+  - @memberjunction/ai@2.57.0
+  - @memberjunction/ai-openai@2.57.0
+
 ## 2.56.0
 
 ### Patch Changes