+++ packages/ExternalChangeDetection/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log - @memberjunction/external-change-detection
 
+## 2.57.0
+
+### Patch Changes
+
+- Updated dependencies [0ba485f]
+  - @memberjunction/core@2.57.0
+  - @memberjunction/core-entities@2.57.0
+  - @memberjunction/global@2.57.0
+  - @memberjunction/sqlserver-dataprovider@2.57.0
+
 ## 2.56.0
 
 ### Patch Changes