+++ packages/MJCore/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log - @memberjunction/core
 
+## 2.57.0
+
+### Minor Changes
+
+- 0ba485f: various bug fixes
+
+### Patch Changes
+
+- Updated dependencies [0ba485f]
+  - @memberjunction/global@2.57.0
+
 ## 2.56.0
 
 ### Patch Changes