+++ packages/MJCoreEntities/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log - @memberjunction/core-entities
 
+## 2.57.0
+
+### Minor Changes
+
+- 0ba485f: various bug fixes
+
+### Patch Changes
+
+- Updated dependencies [0ba485f]
+  - @memberjunction/core@2.57.0
+  - @memberjunction/global@2.57.0
+
 ## 2.56.0
 
 ### Minor Changes