+++ packages/MJCoreEntitiesServer/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @memberjunction/core-entities-server
 
+## 2.57.0
+
+### Minor Changes
+
+- 0ba485f: various bug fixes
+
+### Patch Changes
+
+- Updated dependencies [0ba485f]
+  - @memberjunction/core@2.57.0
+  - @memberjunction/core-entities@2.57.0
+  - @memberjunction/global@2.57.0
+  - @memberjunction/sqlserver-dataprovider@2.57.0
+  - @memberjunction/aiengine@2.57.0
+  - @memberjunction/ai-vector-dupe@2.57.0
+  - @memberjunction/skip-types@2.57.0
+
 ## 2.56.0
 
 ### Minor Changes