+++ packages/SQLServerDataProvider/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log - @memberjunction/sqlserver-dataprovider
 
+## 2.57.0
+
+### Minor Changes
+
+- 0ba485f: various bug fixes
+
+### Patch Changes
+
+- Updated dependencies [0ba485f]
+  - @memberjunction/core@2.57.0
+  - @memberjunction/core-entities@2.57.0
+  - @memberjunction/global@2.57.0
+  - @memberjunction/aiengine@2.57.0
+  - @memberjunction/ai-vector-dupe@2.57.0
+  - @memberjunction/actions@2.57.0
+  - @memberjunction/queue@2.57.0
+  - @memberjunction/ai@2.57.0
+
 ## 2.56.0
 
 ### Patch Changes